Move all settings chooser button code to it's own file
There are at least 2 benefits to doing so:
1. The effects and tween chooser button code had a great deal of duplication with the BaseChooserButton class in SettingsWidgets.py, but they couldn't be subclasses of BaseChooserButton because SettingsWidgets.py was already importing EffectsWidgets.py. With BaseChooserWidget in the same file, this is no longer an issue, which allows simplification of code and makes it easier to maintain
2. SettingsWidgets.py is getting very large. This change makes it easier to maintain the settings widgets.
